Fix VoicemailPlaybackTests.
There appears to be two issues which were borking this, although I'm not sure why
there was the change (maybe it's been around a long time).

A single instance of presenter is retained over time, which is no good for our case
where we're creating different FakeAsyncTaskExecutors each time, so added a method
for testing where we can clear the instance.

Secondly, the method checking the text views didn't work anymore. Not sure why, but
I added a hook into the state text so the tests can continue to check for that.

Bug: 23640774
Change-Id: Ie729627f3bb4ee08476d5ad0198e43cfea72ce5c