Display duration and prepare voicemail if voicemail is downloaded.
We should always know what the duration of the voicemail is even before
it is downloaded. As such, when a voicemail is expanded, set the voicemail
duration. For a downloaded voicemail, we can go even further and prepare
the audio file when the card is expanded. This means the voicemail can
be played immediately when the user clicks "play" instead of having to
wait for the media player to buffer each time.

Bug: 23566924

Change-Id: I0ce7570eab72b0bf079d3eb8cd10b71981576a9f