Protect against more Telecom privileged operations
Make sure that all invocations of TelecomManager methods that
could possibly require a permission are protected by a permission
check.

Some of these are overcautious - for example, the UI should never
show the option to return to a call (READ_PHONE_STATE) if we didn't
detect an active call (READ_PHONE_STATE) in the first place, so
it is not strictly necessary to protect against the former. But not
crashing is the most preferable of all options.

Bug: 20266292

Change-Id: Id91dd16e34320a5e607f91dbce9a4296025eeaaf