UI for InCall context for outgoing calls to a contact.
For outgoing calls, we want to display recent messages exchanged between
the caller and the contact for additional context. This CL is just the
UI components. The values are also not finalized, but getting the UI in
place will be helpful for inserting the values when the logic is in the
place.

Bug: 23351559

Change-Id: I11f504cfd4ee7ea488038c849239117beff9a5de