PhoneStateListener & ContentObserver for blocked calls.
Add a phone state listener to attempt to silence the ringer earlier.

Add ContentObserver which watches the call log. There is an instance
created for each blocked calls. It will watch the call log for changes
pertaining to a set of blocked numbers/times it keeps track of. If
there is a call log entry which matches with that number and time,
mark it as blocked.

Bug: 24341350
Bug: 23944005
Change-Id: I5c40f722d777355140bf98f897d208bd0330e09c