Dismiss filter dialog on rotation.
This simplifies addressing a number of rotation cases by
eliminating them completely. Otherwise, we need to worry
about updating references, such as having the right
fragment or activity reference to be able to provide the
correct parent view for the snackbar. This is particularly
complicated in some cases in Dialer because we have a
fragment => adapter => viewholder hierarchy to pass through.

Bug: 24638328
Change-Id: Ic8a5d2c3ca35742f5bb0986d77ae3b8d5f44e4dd