Don't set FilteredNumberAsyncQueryHandler.
Rather than creating single instances and setting them or passing
them around, create them when we need them. This won't happen too
often so instantiating the objects shouldn't be too much of a
burden on the system.

Instances of AsyncQueryHandler all end up running tasks off
the same thread, so there should not be a negative consequence of
spinning off more threads.

Bug: 24638328
Change-Id: If53ce4f1fd83736c330c1ff02946ed45da575796