Combine blocked number fragments into Activity.
Mocks indicate that after adding, the user should be returned to
the management UI and the snackbar should be shown there. This is
kind of pain to do across fragments in different activities. To
help address this, make one Activity the parent of both the blocked
number manage and add/search UI. This also makes sense in terms of
logical organization.

Bug: 24134038
Change-Id: I682342a73c964a5d4eb90469f10d6584b85400f1