Handle child number changes after the call starts.
The child number display functionality assumed that the child number for
a call would only bet set at the start of the call.  This change removes
that assumption and supports changes to the child number at any point
during the call by adding a new listener to the InCall Call List.  It
appears there are some instances in reality where the child number can
come in after the start of a call (delayed only slightly, but enough to
prevent the number from showing up).

Bug: 24585039
Change-Id: I23148e8c4265f1bc16ce563f2919e9b3eb71784f