Correctly set query string for new SmartDialSearchFragment instance
If DialtactsActivity is created for the first time by launching a
DIAL intent, the SmartDialSearchFragment is not correctly populated
with the correct number for the "Create new contact" and "Add to
contact" shortcuts since it is not attached yet when the callback is
fired. Fix this by persisting the number and setting it on the
fragment when it is attached.

Bug: 24358325
Change-Id: I24d4603376e739a7c9368a9b115109ba23753585