Fix up blocked photo scenarios.
+ Show blocked drawable in CallDetailActivity. Rearrange logic in
the activity to facilitate updating the contact photo and block
item action in accordance to blocked state.
+ Fix bug in FilterNumberDialogFragment where callback was not
always invoked because of final/scope issues where it was possible
for the callback to be unintentionally null and not invoked.
+ Clear blocked id cache when pausing the call log adapter, so that
after changing the blocked state in the call detail activity the
changes will be reflected in the call log.

Bug: 24871853
Change-Id: I1d58b1e0c222ead90fa7b6c30a95bc3254a14776