Move secondary call information to the bottom, clear call context.
The secondary call information was obscuring the call context info
because it was incorrectly being placed right under the primary call
info. This is a bug because it should placed under the FAB at the bottom
of the screen.
Also hide the call context info when a new call comes in that does not
have caller contex.

Bug: 25222360
Change-Id: I7601ce7e7e2609ac046ea104651ac50d863a86bc