Backporting calls to TelecomManager#silenceRinger
The method was a system API pre-M, but a public method in M. Added a
compat method which checks if the method is callable. In the event that
the method can't be called, the compat method is a noop - this matches
the functionality from the L version of the phone.

Bug=25776171
Change-Id: I4a9fc19bb2bf2749d697da2b8643601ebd1eb4a8