Showing badge icon in incall ui if it is a work call
Work call means the call is placed through work phone account
or it is an outgoing call initiated from work dialer.
Please notice that both personal and work call is handled by personal
in-call UI, that's why we need to add a badge to indicate it.

There is another scenario we need to add badge, ie. having a work
contact caller, and it will be handled in ag/836110.


BUG=26022174

Change-Id: I96e72068f70674dbeabfbbf7bfffc6c15c3203bf