Use DATA.CONTACT_ID when lookup SIP call
It is actually a naming mistake in framework, the column store contact id
of normal query is _id, but that of sip query is contact_id.
It is an old issue and we can't do much in the framework side.

So, we need to use the correct projection when it is a SIP call.

Bug:27143980
Bug:27141566

(cherry picked from commit 328f75ff518bf7d8863c6cf4446e26636e05b297)

Change-Id: I0551c308072892ef13c8cb14ea4a4ab37ffa6f83