Check content before requesting a voicemail audio
Formerly dialer only checks for content when the view is loaded, will
only acknowledge the new content status if the fetch succeeded, and
will send a intent every time it thinks there is no content.

Suppose the server took really long to finish the fetch, but eventually
succeeds, it will cause the voicemail fetch to fail every time.

After this fix before a fetch intent is sent, content availability will
be checked first.

The following step is taken to simulate a slow fetching server:
1.disable prefetch in telephony
2.disable the content fetch on dialer load
3.set fetch timeout to 1ms
Verified a similar behavior to b/27388041. Voicemail load fails even if
the content is already there.

This fix was then applied. Voicemail fail for the first time as it should
(timeout), but plays normally after the second time.

Bug:27412247
Change-Id: I184e04fb88772f9006cf6db0fe50d6e773a45d4a