Add default phone account fallback for voicemail notification
Sometimes the accountComponentName and accountId field of the new
voicemail will not be populated by the voicemail source, so a
PhoneAccountHandle used to retrieve the voicemail ringtone can not be
created. Previously we fallback to the default notification sound.

In this CL we will attempt to get the voicemail ringtone from the
default phone account before falling back to notification sound.

Bug:21699149
Change-Id: Ia1c09f5aa32c000cca245c669304cacf4f792c01