Updating logic for isBlockedNumber
In the framework blocking, both the e164 and non-e164 formats of a
number can be blocked at the same time; these two formats will still
have the same e164 value. Since the original code expected the e164
column to be unique, in this situation the Dialer thought the number
was not blocked, resulting in a crash when attempting to block the
number a second time.

This CL changes the logic so multiple rows with the same e164 value
are expected.

Change-Id: I90995046df74b8ec2198eab1ef56a13f20dc1398
Fixes: 27724726