Clear voicemail status text when loading succeeded
In ag/874775, mStateText.setText(null) is removed from
VoicemailPlaybackLayout#setClipPosition(). This caused the error in
mStateText not to be cleared after the voicemail is sucessfully loaded.

This CL adds a setSuccess() method to the layout that clears the message,
which will be called from onPrepared() once the voicemail is loaded.

Fixes:27770842
Change-Id: Ic42e40b99d36f817532dec780b63311c1967300c