Show "manage conference" for video calls.
The existing VT implementation assumed that video conference calling is
not supported, and hence that the "manage conference" option could not
be shown for a video call.  This was based on early requirements, but since
the video conference calling has become a requirement for some carriers.

Removed the video call check in "shouldShowManageConference" and replaced
with a check for fullscreen mode.  This ensures that when a video call is
in fullscreen immersive mode, the "manage conference" banner is hidden
along with the call card.

Also added code in "onFullscreenModeChanged" which hides or shows the
manage conference option as the user enters and exists full screen mode.

Bug: 27616422
Change-Id: Ie23ebe2a0aed2e46166468fa3e833072c0e7a34a