Performing auto-migration when there are no blocked nums
The previous behavior was such that users that had no numbers in the
Dialer block list still had to go through steps to migrate to the
framework blocking solution, even though there was nothing to
migrate. This changes add logic to check if the blocked list is
empty, if that's the cause the user is automatically switched to the
framework blocking solution. This check is performed only once on app
startup.

Fixes: 27704106
Change-Id: I32482003279ef2070c1ebd8b801acf637c48ba8c
(cherry picked from commit ee4e15b40f0d20b5f04f3569efa12c0d26a38645)