Update AOSP Dialer source from internal google3 repository at cl/150392808
Test: Treehugger

Notable this release:
* Reintroduce localization.

This CL updates the AOSP Dialer source with all the changes that have
gone into the private google3 repository. This includes all the
changes from 3/15/2017 to cl/150392808 (3/16/2017).

This goal of these drops is to keep the AOSP source in sync with the
internal google3 repository. Currently these sync are done by hand
with very minor modifications to the internal source code.

See the Android.mk file for list of modifications.

Our current goal is to do frequent drops (daily if possible) and
eventually switched to an automated process.

Change-Id: Ia16d68469c2ca1ff42c7fa31d922bdb78c4e244a