Update AOSP Dialer source from internal google3 repository at cl/151128062
Test: make, treehugger

This CL updates the AOSP Dialer source with all the changes that have
gone into the private google3 repository. This includes all the
changes from cl/150756069 (3/21/2017) to cl/151128062 (3/24/2017).

Notable this release:
- Explicitly enumerate host and target dependencies.
- Update proguard flag references.

This goal of these drops is to keep the AOSP source in sync with the
internal google3 repository. Currently these sync are done by hand
with very minor modifications to the internal source code.
See the Android.mk file for list of modifications.
Our current goal is to do frequent drops (daily if possible) and
eventually switched to an automated process.

Bug: 33210202 36511925

Addresses
33210202 - Proguard support
36511925 - Compiler warnings when building against platform sdk

Change-Id: I448ec3b3f2358886859cf7a4ef76a8fcef3244ae