Prepare for removal of junit and android.test classes from Android API (step 1)
In preparation for removing junit and android.test classes from
the Android API this change adds explicit dependencies on junit
and/or legacy-android-test to ensure that modules will compile
properly once it is removed.

Care was taken to ensure that any targets that depended,
directly or indirectly, on android-support-test did not also
depend on junit or hamcrest as they conflict with the versions
embedded within android-support-test.

The SpannedTestUtils class which is built into the production
Dialer app contains references to JUnit. Rather then embed JUnit in
Dialer it was excluded from the build. It is not actually used at
all anywhere in Dialer app so may even be able to be deleted.

Bug: 30188076
Test: make checkbuild
Change-Id: I37a618deba48d87a1a03d82ad4022233b408a788