Set camera to null when disabling IMS video transmission
From Tyler's investigation:

In the past (Marshmallow days) we used to call setCamera(null) as well
as sending the downgrade request.  Stopping the video transmission is
one thing but I think the modem implementation depends on the camera
being nulled to clean up its resources.

(This CL is a manual cherry pick of cl/157154988 from google3 to
gerrit.)

Bug: 38395481
Change-Id: I7f71c1ffb7194165b8968cd4a033ffea1af443ed