Disable Orange VVM in AOSP dialer
Orange has made a few request that cannot be resolved in the
O timeframe. VVM support for it should be disabled in AOSP so OEMs
won't get it by default without the fixes. These issues will be
resolved in August when O is released and dialer is doing regular
drops in the public.

France is live in prod while Belgium/Luxembourg is experimental
behind a flag.

This CL does not affect GoogleDialer in the system image

Change-Id: Ia7e5c9b5d2411a15453f585bb9eae88aae1bbc00
Fixes: 62536071
Test: manual