Added privacty snackbar for deleting calls with EC data.
When a user deletes a call from the call log, it might not be clear that they
aren't deleting the attachments they made with the call. So we present them
with a snackbar letting them know to open android messages to delete
attachments.

screenshot: http://screen/TSoWmNhjcn4
screenshot: http://screen/vUXShdYpXem
PiperOrigin-RevId: 159631807
Change-Id: Iaad903258160dc789e0c6a524adba56299523636