Fix assert when OmtpMessageSender received an invalid VVM config
b/62666193 is one of the top crashers, the assert that the VVM config is valid has failed. This is likely caused by the PhoneAccountHandle being invalid. While the PhoneAccountHandle must be valid when the broadcast is made, the SIM might have been removed and invalidated the PhoneAccountHandle when the broadcast is still in flight.

In this CL the validity is checked first.

PiperOrigin-RevId: 159737846
Change-Id: Iae5a65978879fd1bfdf726008e8b6463499c99e8